Create requirements.txt
The requirements.txt file is used in Python projects to specify external dependencies (third-party packages) that are required to run the code. Each line in the file lists a package along with its version, which allows others to install all necessary packages in the correct versions with a single command.

This file indicates that the project requires version 4.2.0 of the bpy library. To install the dependencies listed in requirements.txt, someone can run:

pip install -r requirements.txt